# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SEAPEDIA API",
    "version": "1.0.0",
    "description": "Multi-role e-commerce platform API. Supports Admin, Seller, Buyer, and Driver workflows. Protected endpoints require a Bearer JWT token in the Authorization header."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "List products (public)",
        "description": "Paginated catalog. Filters: q, category (slug), sort (newest/price_asc/price_desc/popular), storeId.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "newest",
                "price_asc",
                "price_desc",
                "popular"
              ]
            }
          },
          {
            "name": "storeId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated product list with store info"
          }
        }
      }
    },
    "/products/{slug}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get product by slug (public)",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product detail"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  }
}